Tools for dataset/features manipulation
Imported from tum_veri branch.

The tools allow to put the following dataset to the Freiburg dataset
convention, with the addition of a dataset.txt file
- Convert the microsoft rgbd dataset
- Convert the Freiburg rgbd dataset
- Convert the washington scene dataset

Also, they convert the previous dataset to the leveldb dataset (embed
and resize images), and generate a loop closures dataset containing all
the loop-closures, the positives, the negatives.

It also contains tools to visualize the loop-closure dataset, and plot
the extracted features.