Create ShuffleDataLayer
[IMPORTANT] The working principle of the layer as described below
doesn't work! Refer to more recent commits for a working version!

The shuffle data layer is meant as an input layer for siamese networks.

The shuffle data layer reads a shuffle list containing pairings of
image keys (corresponding to keys in the leveldb dataset). The shuffle
list is then used to load the appropiate data from the dataset.

The data is loaded using two channels:

- On first forward pass, a batch is read from the first channel,
which reads the keys in first column of the shuffle list

- On the next forward pass, a batch is read from the second channel,
  which reads the keys in the second column of the shuffle list

The network definition should only contain one ShuffleDataLayer.
The SplitLayer will split it up between the two sides of the Siamese
network.

Also, this sets the labels for the image pairs in the following way:
label = 1 : image pair is a loop closure
label = 0 otherwise

Wrong LMDB backend code removed.
TODO: support the LMDB backend